transfer training_followup to adoptions
